+++ hosts/common/global/default.nix
@@ -1,8 +1,11 @@
 # Imported by every host.
+{ flake, ... }:
 {
   imports = [
     ./generation.nix
     ./nix.nix
     ./home-manager.nix
   ];
+
+  nixpkgs.overlays = [ flake.outputs.overlays.codex ];
 }